docs: break out llms.txt plugin into own file
